bug #317 code cleaning
git-svn-id: file:///home/svn/shoola-omero/branches@1541 6f2cb1de-eb0d-0410-b157-e593188b5901

+++ OME_2_2/SRC/org/openmicroscopy/shoola/agents/rnd/RenderingAgtCtrl.java
@@ -222,13 +222,13 @@ public class RenderingAgtCtrl
 	/** Forward event to {@link RenderingAgt abstraction}. */
 	public void setChannelWindowStart(int w, int x)
 	{
-		abstraction.setChannelWindowStart(w, (double) x);
+		abstraction.setChannelWindowStart(w, x);
 	}
 	
 	/** Forward event to {@link RenderingAgt abstraction}. */
 	public void setChannelWindowEnd(int w, int x)
 	{
-		abstraction.setChannelWindowEnd(w, (double) x);
+		abstraction.setChannelWindowEnd(w, x);
 	}
 	
 	/** Forward event to {@link RenderingAgt abstraction}. */
