+++ OME_2_2/SRC/org/openmicroscopy/shoola/env/rnd/data/FloatBEConverter.java
@@ -85,7 +85,7 @@ public class FloatBEConverter
 
 		}
 		
-		return (double) Float.intBitsToFloat(r);
+		return Float.intBitsToFloat(r);
 	}
 
 }
