+++ OME_2_2/SRC/org/openmicroscopy/shoola/env/rnd/data/FloatLEConverter.java
@@ -84,7 +84,7 @@ public class FloatLEConverter
 
 		}
 		
-		return (double) Float.intBitsToFloat(r);
+		return Float.intBitsToFloat(r);
 	}
 
 }
